+++ src/cyclops/CyclicCoordinateDescent.cpp
@@ -1084,8 +1084,8 @@ void CyclicCoordinateDescent::findMode(
 		modelSpecifics.setPriorParams(temp);
 		//modelSpecifics.resetBeta();
 	}
-	
-	auto cycle = [this,&lastObjFunc,&lastObjFuncVec,&iteration,algorithmType,&allDelta,&doItAll] {
+
+	auto cycle = [this,&iteration,algorithmType,&allDelta,&doItAll] {
 /*
 		if (iteration%10==0) {
 			std::cout<<"iteration " << iteration << " ";
@@ -1172,7 +1172,7 @@ void CyclicCoordinateDescent::findMode(
 						if (delta != 0.0) {
 							sufficientStatisticsKnown = false;
 							updateSufficientStatistics(delta, index);
-						}	
+						}
 				}
 	    			log(index);
 	    		}
@@ -1813,6 +1813,17 @@ void CyclicCoordinateDescent::turnOnSyncCV(int foldToCompute) {
 void CyclicCoordinateDescent::turnOffSyncCV() {
 	syncCV = false;
 	modelSpecifics.turnOffSyncCV();
+}
+
+void CyclicCoordinateDescent::getSchoenfeldResiduals(const IdType index,
+                                                     std::vector<double>& residuals,
+                                                     std::vector<double>& times) {
+
+    checkAllLazyFlags();
+
+    modelSpecifics.computeSchoenfeldResiduals(index,
+                                              residuals, times,
+                                              false);
 }
 
 std::vector<double> CyclicCoordinateDescent::getPredictiveLogLikelihood(std::vector<std::vector<double>>& weightsPool) {